修复参数格式错误

+++ service-manager/src/main/java/com/moseeker/servicemanager/web/controller/referral/form/ActivityForm.java
@@ -13,7 +13,7 @@ public class ActivityForm {
     private Integer target;
     private String startTime;
     private String endTime;
-    private Double totalAmount;
+    private Integer totalAmount;
     private Double rangeMin;
     private Double rangeMax;
     private Integer probability;
@@ -69,11 +69,11 @@ public class ActivityForm {
         this.endTime = endTime;
     }
 
-    public Double getTotalAmount() {
+    public Integer getTotalAmount() {
         return totalAmount;
     }
 
-    public void setTotalAmount(Double totalAmount) {
+    public void setTotalAmount(Integer totalAmount) {
         this.totalAmount = totalAmount;
     }
